requirements.txt
appdirs==1.4.3
astroid==2.3.3
bcrypt==3.1.7
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
distlib==0.3.0
dnspython==1.16.0
favicon==0.7.0
filelock==3.0.12
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pipenv==2018.11.26
pycparser==2.20
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.12.0
razorpay==1.2.0
requests==2.23.0
sib-api-v3-sdk==6.0.3
six==1.14.0
soupsieve==2.0
stripe==2.43.0
urllib3==1.25.8
virtualenv==20.0.10
virtualenv-clone==0.5.3
Werkzeug==1.0.0
wrapt==1.11.2
